fix(core): rename set-valildator directory to set-validator
